+++ src/ua.rs
@@ -5,6 +5,7 @@ mod array;
 mod browse_result_mask;
 #[cfg(feature = "mbedtls")]
 mod certificate_format;
+mod certificate_verification;
 mod client;
 mod client_config;
 mod continuation_point;
@@ -15,6 +16,7 @@ mod logger;
 mod monitored_item_id;
 mod node_class_mask;
 mod secure_channel_state;
+mod security_level;
 mod server;
 mod server_config;
 mod session_state;
@@ -28,6 +30,7 @@ pub use self::{
     access_level::AccessLevel,
     array::Array,
     browse_result_mask::BrowseResultMask,
+    certificate_verification::CertificateVerification,
     client::{Client, ClientState},
     continuation_point::ContinuationPoint,
     data_types::*,
@@ -36,6 +39,7 @@ pub use self::{
     monitored_item_id::MonitoredItemId,
     node_class_mask::NodeClassMask,
     secure_channel_state::SecureChannelState,
+    security_level::SecurityLevel,
     server::Server,
     session_state::SessionState,
     specified_attributes::SpecifiedAttributes,