+++ src/ua/data_types/message_security_mode.rs
@@ -1,0 +1,7 @@
+crate::data_type!(MessageSecurityMode, UInt32);
+
+crate::enum_variants!(
+    MessageSecurityMode,
+    UA_MessageSecurityMode,
+    [INVALID, NONE, SIGN, SIGNANDENCRYPT]
+);